refactor: move conmpressor functions to infrastructure
