+++ src/libecalc/infrastructure/energy_components/legacy_consumer/component.py
@@ -18,18 +18,6 @@ from libecalc.common.utils.rates import (
     TimeSeriesStreamDayRate,
 )
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.core.consumers.legacy_consumer.consumer_function import (
-    ConsumerFunction,
-    ConsumerFunctionResult,
-)
-from libecalc.core.consumers.legacy_consumer.result_mapper import (
-    get_consumer_system_models,
-    get_operational_settings_results_from_consumer_result,
-    get_single_consumer_models,
-)
-from libecalc.core.consumers.legacy_consumer.system import (
-    ConsumerSystemConsumerFunctionResult,
-)
 from libecalc.core.models.results import CompressorTrainResult
 from libecalc.core.result import ConsumerSystemResult, EcalcModelResult
 from libecalc.core.result.results import (
@@ -40,6 +28,18 @@ from libecalc.core.result.results import (
 )
 from libecalc.expression import Expression
 from libecalc.infrastructure.energy_components.base import BaseConsumer
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function import (
+    ConsumerFunction,
+    ConsumerFunctionResult,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.result_mapper import (
+    get_consumer_system_models,
+    get_operational_settings_results_from_consumer_result,
+    get_single_consumer_models,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.system import (
+    ConsumerSystemConsumerFunctionResult,
+)
 
 
 def get_operational_settings_used_from_consumer_result(
