+++ src/libecalc/infrastructure/energy_components/legacy_consumer/consumer_function/compressor_consumer_function.py
@@ -4,16 +4,6 @@ import numpy as np
 
 from libecalc.common.utils.rates import Rates
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.core.consumers.legacy_consumer.consumer_function import (
-    ConsumerFunction,
-    ConsumerFunctionResult,
-)
-from libecalc.core.consumers.legacy_consumer.consumer_function.utils import (
-    apply_condition,
-    apply_power_loss_factor,
-    get_condition_from_expression,
-    get_power_loss_factor_from_expression,
-)
 from libecalc.core.models.compressor.base import (
     CompressorModel,
     CompressorWithTurbineModel,
@@ -23,6 +13,16 @@ from libecalc.core.models.compressor.train.variable_speed_compressor_train_commo
 )
 from libecalc.core.models.results import CompressorTrainResult
 from libecalc.expression import Expression
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function import (
+    ConsumerFunction,
+    ConsumerFunctionResult,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function.utils import (
+    apply_condition,
+    apply_power_loss_factor,
+    get_condition_from_expression,
+    get_power_loss_factor_from_expression,
+)
 
 
 class CompressorConsumerFunction(ConsumerFunction):
