+++ src/libecalc/infrastructure/energy_components/legacy_consumer/consumer_function/consumer_function.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.core.consumers.legacy_consumer.consumer_function.results import (
+from libecalc.expression import Expression
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function.results import (
     ConsumerFunctionResult,
 )
-from libecalc.expression import Expression
 
 
 class ConsumerFunction(ABC):
