+++ src/libecalc/infrastructure/energy_components/legacy_consumer/consumer_function/consumer_tabular_energy_function.py
@@ -3,22 +3,22 @@ import numpy as np
 from libecalc.common.list.list_utils import array_to_list
 from libecalc.common.utils.rates import Rates
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.core.consumers.legacy_consumer.consumer_function import (
-    ConsumerFunction,
-    ConsumerFunctionResult,
-)
-from libecalc.core.consumers.legacy_consumer.consumer_function.utils import (
-    apply_condition,
-    apply_power_loss_factor,
-    get_condition_from_expression,
-    get_power_loss_factor_from_expression,
-)
 from libecalc.core.models.tabulated import (
     ConsumerTabularEnergyFunction,
     Variable,
     VariableExpression,
 )
 from libecalc.expression import Expression
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function import (
+    ConsumerFunction,
+    ConsumerFunctionResult,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function.utils import (
+    apply_condition,
+    apply_power_loss_factor,
+    get_condition_from_expression,
+    get_power_loss_factor_from_expression,
+)
 
 
 class TabulatedConsumerFunction(ConsumerFunction):
