+++ src/libecalc/infrastructure/energy_components/legacy_consumer/consumer_function/direct_expression_consumer_function.py
@@ -7,18 +7,18 @@ from libecalc.common.list.list_utils import array_to_list
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import Rates, RateType
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.core.consumers.legacy_consumer.consumer_function import (
+from libecalc.core.models.results import EnergyFunctionGenericResult
+from libecalc.expression import Expression
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function import (
     ConsumerFunction,
     ConsumerFunctionResult,
 )
-from libecalc.core.consumers.legacy_consumer.consumer_function.utils import (
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function.utils import (
     apply_condition,
     apply_power_loss_factor,
     get_condition_from_expression,
     get_power_loss_factor_from_expression,
 )
-from libecalc.core.models.results import EnergyFunctionGenericResult
-from libecalc.expression import Expression
 
 
 class DirectExpressionConsumerFunction(ConsumerFunction):
