+++ src/libecalc/infrastructure/energy_components/legacy_consumer/consumer_function/utils.py
