+++ src/libecalc/infrastructure/energy_components/legacy_consumer/consumer_function_mapper/compressor_system_consumer_function.py
@@ -1,17 +1,17 @@
 from typing import Optional
 
-from libecalc.core.consumers.legacy_consumer.system.consumer_function import (
-    CompressorSystemConsumerFunction,
-    ConsumerSystemConsumerFunction,
-)
-from libecalc.core.consumers.legacy_consumer.system.operational_setting import (
-    CompressorSystemOperationalSettingExpressions,
-)
-from libecalc.core.consumers.legacy_consumer.system.types import ConsumerSystemComponent
 from libecalc.core.models.compressor import create_compressor_model
 from libecalc.dto import CompressorSystemConsumerFunction as CompressorSystemConsumerFunctionDTO
 from libecalc.dto import CompressorSystemOperationalSetting, SystemOperationalSetting
 from libecalc.expression import Expression
+from libecalc.infrastructure.energy_components.legacy_consumer.system.consumer_function import (
+    CompressorSystemConsumerFunction,
+    ConsumerSystemConsumerFunction,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.system.operational_setting import (
+    CompressorSystemOperationalSettingExpressions,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.system.types import ConsumerSystemComponent
 
 
 def _map_operational_settings(
