+++ src/libecalc/infrastructure/energy_components/legacy_consumer/consumer_function_mapper/pump_system_consumer_function.py
@@ -1,17 +1,17 @@
 from typing import Optional
 
-from libecalc.core.consumers.legacy_consumer.system.consumer_function import (
-    ConsumerSystemConsumerFunction,
-    PumpSystemConsumerFunction,
-)
-from libecalc.core.consumers.legacy_consumer.system.operational_setting import (
-    PumpSystemOperationalSettingExpressions,
-)
-from libecalc.core.consumers.legacy_consumer.system.types import ConsumerSystemComponent
 from libecalc.core.models.pump import create_pump_model
 from libecalc.dto import PumpSystemConsumerFunction as PumpSystemConsumerFunctionDTO
 from libecalc.dto import PumpSystemOperationalSetting
 from libecalc.expression import Expression
+from libecalc.infrastructure.energy_components.legacy_consumer.system.consumer_function import (
+    ConsumerSystemConsumerFunction,
+    PumpSystemConsumerFunction,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.system.operational_setting import (
+    PumpSystemOperationalSettingExpressions,
+)
+from libecalc.infrastructure.energy_components.legacy_consumer.system.types import ConsumerSystemComponent
 
 from .compressor_system_consumer_function import (
     map_discharge_pressures,
