+++ src/libecalc/infrastructure/energy_components/legacy_consumer/system/operational_setting.py
