+++ src/libecalc/infrastructure/energy_components/legacy_consumer/system/types.py
