+++ tests/libecalc/core/consumers/consumer_function/test_direct_expression_consumer_function.py
@@ -11,7 +11,7 @@ from libecalc.common.temporal_model import TemporalModel
 from libecalc.common.time_utils import Period
 from libecalc.common.variables import VariablesMap
 from libecalc.infrastructure.energy_components.legacy_consumer.component import Consumer
-from libecalc.core.consumers.legacy_consumer.consumer_function.direct_expression_consumer_function import (
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function.direct_expression_consumer_function import (
     DirectExpressionConsumerFunction,
 )
 from libecalc.expression import Expression
