+++ tests/libecalc/core/consumers/system/test_operational_setting.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from libecalc.common.errors.exceptions import EcalcError
-from libecalc.core.consumers.legacy_consumer.system.operational_setting import (
+from libecalc.infrastructure.energy_components.legacy_consumer.system.operational_setting import (
     ConsumerSystemOperationalSetting,
     ConsumerSystemOperationalSettingExpressions,
 )
