+++ tests/libecalc/core/consumers/test_consumer_utils.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from libecalc.common.string.string_utils import generate_id
 from libecalc.infrastructure.energy_components.consumer_system import ConsumerSystem
-from libecalc.core.consumers.legacy_consumer.consumer_function.utils import (
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function.utils import (
     apply_condition,
 )
 from libecalc.infrastructure.components import Crossover
