+++ tests/libecalc/integration/test_all_consumer_with_time_slots_models.py
@@ -10,7 +10,7 @@ from libecalc.common.temporal_model import TemporalModel
 from libecalc.common.time_utils import Period, Periods
 from libecalc.common.variables import VariablesMap
 from libecalc.infrastructure.energy_components.legacy_consumer.component import Consumer
-from libecalc.core.consumers.legacy_consumer.consumer_function_mapper import EnergyModelMapper
+from libecalc.infrastructure.energy_components.legacy_consumer.consumer_function_mapper import EnergyModelMapper
 from libecalc.core.result import CompressorModelResult, GenericModelResult
